Rename *status* contribs for clarity
